requirements.txt
asgiref==3.8.1
Django==5.1.4 #https://pypi.org/project/Django/
djangorestframework==3.15.2 #https://pypi.org/project/djangorestframework/
sqlparse==0.5.3
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8 #https://pypi.org/project/drf-yasg/
inflection==0.5.1
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
